Provide Buffer plugin
Provide Buffer plugin in webpack to use it in extension instead of having the following issue "Buffer is not defined".

Exemple : Using the library "shpjs" in an extension trigger the error. The submobule Mapstore2 provides the Buffer in his webpack. To use it in extensions, we need to provide it in the Extension project.